requirement.txt
python-docx
langchain
openai
sentence-transformers
faiss-cpu
python-dotenv